+++ orchestrator/src/buttercup/orchestrator/competition_api_client/models/types_fuzzing_engine.py
@@ -18,20 +18,19 @@ from enum import Enum
 from typing_extensions import Self
 
 
-class TypesAssessment(str, Enum):
+class TypesFuzzingEngine(str, Enum):
     """
-    TypesAssessment
+    TypesFuzzingEngine
     """
 
     """
     allowed enum values
     """
-    AssessmentCorrect = 'correct'
-    AssessmentIncorrect = 'incorrect'
+    FuzzingEngineLibFuzzer = 'libfuzzer'
 
     @classmethod
     def from_json(cls, json_str: str) -> Self:
-        """Create an instance of TypesAssessment from a JSON string"""
+        """Create an instance of TypesFuzzingEngine from a JSON string"""
         return cls(json.loads(json_str))